invoice_rag/requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=9.5.0
plotly>=5.15.0

# AI/ML dependencies
groq>=0.4.0
pydantic>=2.0.0

# Database dependencies
sqlalchemy>=2.0.0

# Supabase dependencies (for PostgreSQL migration)
supabase>=2.22.2
postgrest>=2.22.2
psycopg2-binary>=2.9.11
realtime>=2.22.2
storage3>=2.22.2
supabase-auth>=2.22.2
supabase-functions>=2.22.2
strenum>=0.4.15

# Environment and utilities
python-dotenv>=1.0.0

# Image processing
opencv-python>=4.8.0

# Additional utilities
requests>=2.31.0

# Telegram dependencies
python-telegram-bot

# Premium feature dependencies
PyJWT>=2.8.0

# Spreadsheet export dependencies
openpyxl>=3.1.0
gspread>=5.12.0
oauth2client>=4.1.3

# WhatsApp Bot dependencies (WAHA + n8n integration)
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Marimo interactive dashboard dependencies
marimo>=0.9.0
flask>=3.0.0
